@@ -6,50 +6,90 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+	"time"
 
 	"github.com/KiloProjects/kilonova"
+	"github.com/jmoiron/sqlx"
 )
 
 const createAttachmentQuery = "INSERT INTO attachments (problem_id, visible, private, name, data) VALUES (?, ?, ?, ?, ?) RETURNING id;"
 
-func (a *DB) CreateAttachment(ctx context.Context, att *kilonova.Attachment) error {
-	if att.ProblemID == 0 || att.Data == nil {
+func (a *DB) CreateAttachment(ctx context.Context, att *kilonova.Attachment, problemID int, data []byte) error {
+	if problemID == 0 || data == nil {
 		return kilonova.ErrMissingRequired
 	}
-	if _, err := a.Attachments(ctx, false, kilonova.AttachmentFilter{ProblemID: &att.ProblemID, Name: &att.Name}); err != nil {
+	if _, err := a.ProblemAttachments(ctx, problemID, &kilonova.AttachmentFilter{Name: &att.Name}); err != nil {
 		return kilonova.ErrAttachmentExists
 	}
 
 	var id int
-	err := a.conn.GetContext(ctx, &id, a.conn.Rebind(createAttachmentQuery), att.ProblemID, att.Visible, att.Private, att.Name, att.Data)
+	err := a.conn.GetContext(ctx, &id, a.conn.Rebind(createAttachmentQuery), problemID, att.Visible, att.Private, att.Name, data)
 	if err == nil {
 		att.ID = id
 	}
 	return err
 }
 
+const selectedAttFields = "id, created_at, visible, private, name, data_size" // Make sure to keep this in sync
+
 func (a *DB) Attachment(ctx context.Context, id int) (*kilonova.Attachment, error) {
-	var attachment kilonova.Attachment
-	err := a.conn.GetContext(ctx, &attachment, a.conn.Rebind("SELECT * FROM attachments WHERE id = ? LIMIT 1"), id)
+	var att dbAttachment
+	err := a.conn.GetContext(ctx, &att, "SELECT "+selectedAttFields+" FROM attachments WHERE id = $1 LIMIT 1", id)
 	if errors.Is(err, sql.ErrNoRows) {
 		return nil, nil
 	}
-	return &attachment, err
+	return internalToAttachment(&att), err
 }
 
-func (a *DB) Attachments(ctx context.Context, getData bool, filter kilonova.AttachmentFilter) ([]*kilonova.Attachment, error) {
-	var attachments []*kilonova.Attachment
-	where, args := attachmentFilterQuery(&filter)
-	toSelect := "*"
-	if !getData {
-		toSelect = "id, created_at, problem_id, visible, private, name, data_size" // Make sure to keep this in sync
+func (a *DB) AttachmentByName(ctx context.Context, problemID int, filename string) (*kilonova.Attachment, error) {
+	var att dbAttachment
+	err := a.conn.GetContext(ctx, &att, "SELECT "+selectedAttFields+" FROM attachments WHERE problem_id = $1 AND name = $2 LIMIT 1", problemID, filename)
+	if errors.Is(err, sql.ErrNoRows) {
+		return nil, nil
 	}
-	query := a.conn.Rebind("SELECT " + toSelect + " FROM attachments WHERE " + strings.Join(where, " AND ") + " ORDER BY name ASC " + FormatLimitOffset(filter.Limit, filter.Offset))
+	return internalToAttachment(&att), err
+}
+
+/*
+func (a *DB) Attachments(ctx context.Context, filter kilonova.AttachmentFilter) ([]*kilonova.Attachment, error) {
+	var attachments []*dbAttachment
+	where, args := attachmentFilterQuery(&filter)
+	query := a.conn.Rebind("SELECT " + selectedAttFields + " FROM attachments WHERE " + strings.Join(where, " AND ") + " ORDER BY name ASC " + FormatLimitOffset(filter.Limit, filter.Offset))
 	err := a.conn.SelectContext(ctx, &attachments, query, args...)
 	if errors.Is(err, sql.ErrNoRows) {
 		return []*kilonova.Attachment{}, nil
 	}
-	return attachments, err
+	return internalToAttachments(attachments), err
+}
+*/
+
+func (a *DB) ProblemAttachments(ctx context.Context, pbid int, filter *kilonova.AttachmentFilter) ([]*kilonova.Attachment, error) {
+	var attachments []*dbAttachment
+	where, args := attachmentFilterQuery(filter)
+	query := a.conn.Rebind("SELECT " + selectedAttFields + " FROM attachments WHERE problem_id = $1 AND " + strings.Join(where, " AND ") + " ORDER BY name ASC " + FormatLimitOffset(filter.Limit, filter.Offset))
+	err := a.conn.SelectContext(ctx, &attachments, query, append([]any{pbid}, args...)...)
+	if errors.Is(err, sql.ErrNoRows) {
+		return []*kilonova.Attachment{}, nil
+	}
+	return internalToAttachments(attachments), err
+}
+
+func (a *DB) AttachmentData(ctx context.Context, id int) ([]byte, error) {
+	var data []byte
+	err := a.conn.GetContext(ctx, &data, "SELECT data FROM attachments WHERE id = $1", id)
+	if errors.Is(err, sql.ErrNoRows) {
+		return []byte{}, nil
+	}
+	return data, err
+}
+
+func (a *DB) AttachmentDataByName(ctx context.Context, problemID int, name string) ([]byte, error) {
+	var data []byte
+	err := a.conn.GetContext(ctx, &data, "SELECT data FROM attachments WHERE problem_id = $1 AND name = $1", problemID, name)
+	if errors.Is(err, sql.ErrNoRows) {
+		return []byte{}, nil
+	}
+	return data, err
 }
 
 const attachmentUpdateStatement = "UPDATE attachments SET %s WHERE id = ?"
@@ -70,22 +110,35 @@ func (a *DB) DeleteAttachment(ctx context.Context, attid int) error {
 	return err
 }
 
-func (a *DB) DeleteAttachments(ctx context.Context, pbid int) error {
-	_, err := a.conn.ExecContext(ctx, a.conn.Rebind("DELETE FROM attachments WHERE problem_id = ?"), pbid)
-	return err
+func (a *DB) DeleteAttachments(ctx context.Context, pbid int, attIDs []int) (int64, error) {
+	query, args, err := sqlx.In("DELETE FROM attachments WHERE problem_id = ? AND id IN (?)", pbid, attIDs)
+	if err != nil {
+		return -1, err
+	}
+	result, err := a.conn.ExecContext(ctx, a.conn.Rebind(query), args...)
+	if err != nil {
+		return -1, err
+	}
+
+	return result.RowsAffected()
 }
 
-func attachmentFilterQuery(filter *kilonova.AttachmentFilter) ([]string, []interface{}) {
-	where, args := []string{"1 = 1"}, []interface{}{}
+func attachmentFilterQuery(filter *kilonova.AttachmentFilter) ([]string, []any) {
+	where, args := []string{"1 = 1"}, []any{}
+	if filter == nil {
+		return where, args
+	}
 	if v := filter.ID; v != nil {
 		where, args = append(where, "id = ?"), append(args, v)
 	}
 	if v := filter.Name; v != nil {
 		where, args = append(where, "name = ?"), append(args, v)
 	}
-	if v := filter.ProblemID; v != nil {
-		where, args = append(where, "problem_id = ?"), append(args, v)
-	}
+	/*
+		if v := filter.ProblemID; v != nil {
+			where, args = append(where, "problem_id = ?"), append(args, v)
+		}
+	*/
 	if v := filter.Visible; v != nil {
 		where, args = append(where, "visible = ?"), append(args, v)
 	}
@@ -95,8 +148,8 @@ func attachmentFilterQuery(filter *kilonova.AttachmentFilter) ([]string, []inter
 	return where, args
 }
 
-func attachmentUpdateQuery(upd *kilonova.AttachmentUpdate) ([]string, []interface{}) {
-	toUpd, args := []string{}, []interface{}{}
+func attachmentUpdateQuery(upd *kilonova.AttachmentUpdate) ([]string, []any) {
+	toUpd, args := []string{}, []any{}
 	if v := upd.Name; v != nil {
 		toUpd, args = append(toUpd, "name = ?"), append(args, v)
 	}
@@ -107,4 +160,35 @@ func attachmentUpdateQuery(upd *kilonova.AttachmentUpdate) ([]string, []interfac
 		toUpd, args = append(toUpd, "private = ?"), append(args, v)
 	}
 	return toUpd, args
+}
+
+type dbAttachment struct {
+	ID        int       `db:"id"`
+	CreatedAt time.Time `db:"created_at"`
+	Visible   bool      `db:"visible"`
+	Private   bool      `db:"private"`
+
+	Name string `db:"name"`
+	Size int    `db:"data_size"`
+	//Data []byte `db:"data"`
+}
+
+// TODO: Test
+func internalToAttachments(att []*dbAttachment) []*kilonova.Attachment {
+	return mapper(att, internalToAttachment)
+}
+
+func internalToAttachment(att *dbAttachment) *kilonova.Attachment {
+	if att == nil {
+		return nil
+	}
+	return &kilonova.Attachment{
+		ID:        att.ID,
+		CreatedAt: att.CreatedAt,
+		Visible:   att.Visible,
+		Private:   att.Private,
+
+		Name: att.Name,
+		Size: att.Size,
+	}
 }
