@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/KiloProjects/kilonova"
-	"github.com/jackc/pgtype"
 	"go.uber.org/zap"
 )
 
@@ -19,7 +18,11 @@ func (s *DB) ProblemList(ctx context.Context, id int) (*kilonova.ProblemList, er
 	if errors.Is(err, sql.ErrNoRows) {
 		return nil, nil
 	}
-	return internalToPbList(&pblist), err
+	if err != nil {
+		return nil, err
+	}
+
+	return s.internalToPbList(ctx, &pblist)
 }
 
 func (s *DB) ProblemLists(ctx context.Context, filter kilonova.ProblemListFilter) ([]*kilonova.ProblemList, error) {
@@ -35,30 +38,32 @@ func (s *DB) ProblemLists(ctx context.Context, filter kilonova.ProblemListFilter
 
 	outLists := make([]*kilonova.ProblemList, 0, len(lists))
 	for _, el := range lists {
-		outLists = append(outLists, internalToPbList(el))
+		pblist, err := s.internalToPbList(ctx, el)
+		if err != nil {
+			zap.S().Warn(err)
+			continue
+		}
+		outLists = append(outLists, pblist)
 	}
 	return outLists, err
 }
 
-const createProblemListQuery = "INSERT INTO problem_lists (author_id, title, description, list) VALUES (?, ?, ?, ?) RETURNING id;"
+const createProblemListQuery = "INSERT INTO problem_lists (author_id, title, description) VALUES (?, ?, ?) RETURNING id;"
 
 func (s *DB) CreateProblemList(ctx context.Context, list *kilonova.ProblemList) error {
 	if list.AuthorID == 0 {
 		return kilonova.ErrMissingRequired
 	}
-	// Convert subtask list to array
-	arr := pgtype.Int4Array{}
-	if err := arr.Set(list.List); err != nil {
-		zap.S().Warn("Wtf", err)
-		return err
-	}
 	// Do insertion
 	var id int
-	err := s.conn.GetContext(ctx, &id, s.conn.Rebind(createProblemListQuery), list.AuthorID, list.Title, list.Description, arr)
-	if err == nil {
-		list.ID = id
+	err := s.conn.GetContext(ctx, &id, s.conn.Rebind(createProblemListQuery), list.AuthorID, list.Title, list.Description)
+	if err != nil {
+		return err
 	}
-	return err
+	list.ID = id
+
+	// Add problems
+	return s.UpdateProblemListProblems(ctx, list.ID, list.List)
 }
 
 func (s *DB) UpdateProblemList(ctx context.Context, id int, upd kilonova.ProblemListUpdate) error {
@@ -77,8 +82,30 @@ func (s *DB) DeleteProblemList(ctx context.Context, id int) error {
 	return err
 }
 
-func pblistFilterQuery(filter *kilonova.ProblemListFilter) ([]string, []interface{}) {
-	where, args := []string{"1 = 1"}, []interface{}{}
+func (s *DB) UpdateProblemListProblems(ctx context.Context, id int, problemIDs []int) error {
+	tx, err := s.conn.BeginTxx(ctx, nil)
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+
+	// Naively delete all associations, then add them back
+	if _, err := tx.ExecContext(ctx, s.conn.Rebind("DELETE FROM problem_list_problems WHERE pblist_id = ?"), id); err != nil {
+		return err
+	}
+
+	for _, pbid := range problemIDs {
+		if _, err := tx.ExecContext(ctx, s.conn.Rebind("INSERT INTO problem_list_problems (pblist_id, problem_id) VALUES (?, ?)"), id, pbid); err != nil {
+			zap.S().Warn(err)
+			return err
+		}
+	}
+
+	return tx.Commit()
+}
+
+func pblistFilterQuery(filter *kilonova.ProblemListFilter) ([]string, []any) {
+	where, args := []string{"1 = 1"}, []any{}
 	if v := filter.ID; v != nil {
 		where, args = append(where, "id = ?"), append(args, v)
 	}
@@ -89,19 +116,16 @@ func pblistFilterQuery(filter *kilonova.ProblemListFilter) ([]string, []interfac
 	return where, args
 }
 
-func pblistUpdateQuery(upd *kilonova.ProblemListUpdate) ([]string, []interface{}) {
-	toUpd, args := []string{}, []interface{}{}
+func pblistUpdateQuery(upd *kilonova.ProblemListUpdate) ([]string, []any) {
+	toUpd, args := []string{}, []any{}
 	if v := upd.AuthorID; v != nil {
 		toUpd, args = append(toUpd, "author_id = ?"), append(args, v)
 	}
-	if v := upd.Title; v != nil {
+	if v := upd.Title; v != nil && len(*v) > 0 {
 		toUpd, args = append(toUpd, "title = ?"), append(args, v)
 	}
 	if v := upd.Description; v != nil {
 		toUpd, args = append(toUpd, "description = ?"), append(args, v)
-	}
-	if v := upd.List; v != nil {
-		toUpd, args = append(toUpd, "list = ?"), append(args, kilonova.SerializeIntList(v))
 	}
 	return toUpd, args
 }
@@ -112,19 +136,26 @@ type pblist struct {
 	AuthorID    int       `db:"author_id"`
 	Title       string
 	Description string
-	List        pgtype.Int8Array
 }
 
-func internalToPbList(list *pblist) *kilonova.ProblemList {
-	ids := make([]int, 0, len(list.List.Elements))
-	for _, id := range list.List.Elements {
-		ids = append(ids, int(id.Int))
+func (s *DB) internalToPbList(ctx context.Context, list *pblist) (*kilonova.ProblemList, error) {
+
+	var ids []int
+	err := s.conn.SelectContext(ctx, &ids, s.conn.Rebind("SELECT problem_id FROM problem_list_problems WHERE pblist_id = ? ORDER BY problem_id ASC"), list.ID)
+	if errors.Is(err, sql.ErrNoRows) {
+		ids = []int{}
+	} else if err != nil {
+		return nil, err
 	}
+	if len(ids) == 0 {
+		ids = []int{}
+	}
+
 	return &kilonova.ProblemList{
 		ID:          list.ID,
 		CreatedAt:   list.CreatedAt,
 		Title:       list.Title,
 		Description: list.Description,
 		List:        ids,
-	}
+	}, nil
 }
