@@ -8,6 +8,8 @@ import (
 	"strings"
 
 	"github.com/KiloProjects/kilonova"
+	"github.com/KiloProjects/kilonova/eval"
+	"github.com/jmoiron/sqlx"
 	"go.uber.org/zap"
 )
 
@@ -37,9 +39,6 @@ func (s *DB) Submissions(ctx context.Context, filter kilonova.SubmissionFilter) 
 func (s *DB) CountSubmissions(ctx context.Context, filter kilonova.SubmissionFilter) (int, error) {
 	where, args := subFilterQuery(&filter)
 	query := "SELECT COUNT(id) FROM submissions WHERE " + strings.Join(where, " AND ")
-	if len(where) == 0 {
-		strings.ReplaceAll(query, "WHERE", "")
-	}
 	var cnt int
 	query = s.conn.Rebind(query)
 	err := s.conn.GetContext(ctx, &cnt, query, args...)
@@ -48,26 +47,27 @@ func (s *DB) CountSubmissions(ctx context.Context, filter kilonova.SubmissionFil
 
 const createSubQuery = "INSERT INTO submissions (user_id, problem_id, language, code) VALUES (?, ?, ?, ?) RETURNING id;"
 
-func (s *DB) CreateSubmission(ctx context.Context, sub *kilonova.Submission) error {
-	if sub.UserID == 0 || sub.ProblemID == 0 || sub.Language == "" || sub.Code == "" {
-		return kilonova.ErrMissingRequired
+func (s *DB) CreateSubmission(ctx context.Context, authorID int, problem *kilonova.Problem, language eval.Language, code string) (int, error) {
+	if authorID <= 0 || problem == nil || language.InternalName == "" || code == "" {
+		return -1, kilonova.ErrMissingRequired
 	}
 	var id int
-	err := s.conn.GetContext(ctx, &id, s.conn.Rebind(createSubQuery), sub.UserID, sub.ProblemID, sub.Language, sub.Code)
-	if err == nil {
-		sub.ID = id
-	}
-	return err
+	err := s.conn.GetContext(ctx, &id, s.conn.Rebind(createSubQuery), authorID, problem.ID, language.InternalName, code)
+	return id, err
 }
 
 func (s *DB) UpdateSubmission(ctx context.Context, id int, upd kilonova.SubmissionUpdate) error {
+	return updateSubmission(s.conn, ctx, id, upd)
+}
+
+func updateSubmission(tx sqlx.ExecerContext, ctx context.Context, id int, upd kilonova.SubmissionUpdate) error {
 	toUpd, args := subUpdateQuery(&upd)
 	if len(toUpd) == 0 {
 		return kilonova.ErrNoUpdates
 	}
 	args = append(args, id)
-	query := s.conn.Rebind(fmt.Sprintf(`UPDATE submissions SET %s WHERE id = ?;`, strings.Join(toUpd, ", ")))
-	_, err := s.conn.ExecContext(ctx, query, args...)
+	query := sqlx.Rebind(sqlx.BindType("pgx"), fmt.Sprintf(`UPDATE submissions SET %s WHERE id = ?;`, strings.Join(toUpd, ", ")))
+	_, err := tx.ExecContext(ctx, query, args...)
 	return err
 }
 
@@ -105,7 +105,7 @@ LIMIT 1;`), userid, problemid)
 }
 
 func (s *DB) MaxScores(ctx context.Context, userid int, pbids []int) map[int]int {
-	if pbids == nil || len(pbids) == 0 {
+	if len(pbids) == 0 {
 		return nil
 	}
 
@@ -120,7 +120,7 @@ func (s *DB) MaxScores(ctx context.Context, userid int, pbids []int) map[int]int
 	}
 	inClause += ")"
 
-	args := make([]interface{}, 0, len(pbids))
+	args := make([]any, 0, len(pbids))
 	for _, id := range pbids {
 		args = append(args, id)
 	}
@@ -139,17 +139,14 @@ func (s *DB) MaxScores(ctx context.Context, userid int, pbids []int) map[int]int
 	return rez
 }
 
-func (s *DB) SolvedProblems(ctx context.Context, userid int) ([]int, error) {
+func (s *DB) solvedProblems(ctx context.Context, userid int) ([]int, error) {
 	var pbs []int
-	err := s.conn.SelectContext(ctx, &pbs, s.conn.Rebind(`SELECT problem_id FROM submissions
-WHERE score = 100 AND user_id = ?
-GROUP BY problem_id
-ORDER BY problem_id;`), userid)
+	err := s.conn.SelectContext(ctx, &pbs, `SELECT problem_id FROM submissions WHERE score = 100 AND user_id = $1 GROUP BY problem_id ORDER BY problem_id;`, userid)
 	return pbs, err
 }
 
-func subFilterQuery(filter *kilonova.SubmissionFilter) ([]string, []interface{}) {
-	where, args := []string{"1 = 1"}, []interface{}{}
+func subFilterQuery(filter *kilonova.SubmissionFilter) ([]string, []any) {
+	where, args := []string{"1 = 1"}, []any{}
 	if v := filter.ID; v != nil {
 		where, args = append(where, "id = ?"), append(args, v)
 	}
@@ -173,18 +170,12 @@ func subFilterQuery(filter *kilonova.SubmissionFilter) ([]string, []interface{})
 	if v := filter.CompileError; v != nil {
 		where, args = append(where, "compile_error = ?"), append(args, v)
 	}
-	if v := filter.Visible; v != nil {
-		where, args = append(where, "visible = ?"), append(args, v)
-	}
-	if v := filter.Quality; v != nil {
-		where, args = append(where, "quality = ?"), append(args, v)
-	}
 
 	return where, args
 }
 
-func subUpdateQuery(upd *kilonova.SubmissionUpdate) ([]string, []interface{}) {
-	toUpd, args := []string{}, []interface{}{}
+func subUpdateQuery(upd *kilonova.SubmissionUpdate) ([]string, []any) {
+	toUpd, args := []string{}, []any{}
 	if v := upd.Status; v != kilonova.StatusNone {
 		toUpd, args = append(toUpd, "status = ?"), append(args, v)
 	}
@@ -206,13 +197,6 @@ func subUpdateQuery(upd *kilonova.SubmissionUpdate) ([]string, []interface{}) {
 		toUpd, args = append(toUpd, "max_memory = ?"), append(args, v)
 	}
 
-	if v := upd.Visible; v != nil {
-		toUpd, args = append(toUpd, "visible = ?"), append(args, v)
-	}
-	if v := upd.Quality; v != nil {
-		toUpd, args = append(toUpd, "quality = ?"), append(args, v)
-	}
-
 	return toUpd, args
 }
 
@@ -231,4 +215,37 @@ func getSubmissionOrdering(ordering string, ascending bool) string {
 	default:
 		return "ORDER BY id" + ord
 	}
+}
+
+// FetchGraderSubmission returns a waiting submission that is not locked. Note that it must be closed
+func (s *DB) FetchGraderSubmission(ctx context.Context) (eval.GraderSubmission, error) {
+	tx, err := s.conn.BeginTxx(ctx, nil)
+	if err != nil {
+		return nil, err
+	}
+	var sub kilonova.Submission
+	err = tx.GetContext(ctx, &sub, "SELECT * FROM submissions WHERE status = 'waiting' ORDER BY id DESC LIMIT 1 FOR UPDATE SKIP LOCKED")
+	if errors.Is(err, sql.ErrNoRows) {
+		return nil, tx.Rollback()
+	}
+	return &graderSub{ctx, &sub, tx}, nil
+}
+
+type graderSub struct {
+	ctx context.Context
+	sub *kilonova.Submission
+	tx  *sqlx.Tx
+}
+
+func (g *graderSub) Submission() *kilonova.Submission {
+	return g.sub
+}
+
+func (g *graderSub) Update(sub kilonova.SubmissionUpdate) error {
+	return updateSubmission(g.tx, g.ctx, g.sub.ID, sub)
+}
+
+func (g *graderSub) Close() error {
+	defer g.tx.Rollback()
+	return g.tx.Commit()
 }
