@@ -1,9 +1,7 @@
 package api
 
 import (
-	"io"
 	"net/http"
-	"strconv"
 
 	"github.com/KiloProjects/kilonova"
 	"github.com/KiloProjects/kilonova/internal/util"
@@ -21,11 +19,11 @@ func (s *API) maxScore(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if args.UserID == 0 {
-		if util.User(r) == nil {
+		if util.UserBrief(r) == nil {
 			errorData(w, "No user specified", 400)
 			return
 		}
-		args.UserID = util.User(r).ID
+		args.UserID = util.UserBrief(r).ID
 	}
 
 	if args.ProblemID == 0 {
@@ -33,58 +31,39 @@ func (s *API) maxScore(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	returnData(w, s.db.MaxScore(r.Context(), args.UserID, args.ProblemID))
+	returnData(w, s.base.MaxScore(r.Context(), args.UserID, args.ProblemID))
 }
 
 func (s *API) deleteProblem(w http.ResponseWriter, r *http.Request) {
-	r.ParseForm()
-	if err := s.db.DeleteProblem(r.Context(), util.Problem(r).ID); err != nil {
-		errorData(w, err, 500)
+	if err := s.base.DeleteProblem(r.Context(), util.Problem(r).ID); err != nil {
+		err.WriteError(w)
 		return
 	}
 	returnData(w, "Deleted problem")
 }
 
 // initProblem assigns an ID for the problem
-// TODO: Move most stuff to logic
 func (s *API) initProblem(w http.ResponseWriter, r *http.Request) {
-	title := r.FormValue("title")
-	if title == "" {
-		errorData(w, "Title not provided", http.StatusBadRequest)
+	r.ParseForm()
+	var args struct {
+		Title        string `json:"title"`
+		ConsoleInput bool   `json:"consoleInput"`
+	}
+	if err := decoder.Decode(&args, r.Form); err != nil {
+		errorData(w, err, 400)
 		return
 	}
 
-	cistr := r.FormValue("consoleInput")
-	var consoleInput bool
-	if cistr != "" {
-		ci, err := strconv.ParseBool(cistr)
-		if err != nil {
-			errorData(w, "Invalid `consoleInput` form value", http.StatusBadRequest)
-			return
-		}
-		consoleInput = ci
-	}
-
-	pb, err := s.db.Problems(r.Context(), kilonova.ProblemFilter{Name: &title})
-	if len(pb) > 0 || err != nil {
-		errorData(w, "Problem with specified title already exists in DB", http.StatusBadRequest)
+	id, err := s.base.CreateProblem(r.Context(), args.Title, util.UserBrief(r), args.ConsoleInput)
+	if err != nil {
+		err.WriteError(w)
 		return
 	}
 
-	var problem kilonova.Problem
-	problem.Name = title
-	problem.AuthorID = util.User(r).ID
-	problem.ConsoleInput = consoleInput
-	if err := s.db.CreateProblem(r.Context(), &problem); err != nil {
-		errorData(w, err, 500)
-		return
-	}
-
-	returnData(w, problem.ID)
+	returnData(w, id)
 }
 
 // getProblems returns all the problems from the DB matching a filter
-// TODO: Pagination
 func (s *API) getProblems(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm()
 	var args kilonova.ProblemFilter
@@ -93,139 +72,25 @@ func (s *API) getProblems(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var id int
-	if util.User(r) != nil {
-		id = util.User(r).ID
-		if util.User(r).Admin {
-			id = -1
-		}
-	}
-	args.LookingUserID = &id
+	args.LookingUser = util.UserBrief(r)
 
-	problems, err := s.db.Problems(r.Context(), args)
+	problems, err := s.base.Problems(r.Context(), args)
 	if err != nil {
-		errorData(w, http.StatusText(500), 500)
+		err.WriteError(w)
 		return
 	}
 	returnData(w, problems)
 }
 
-// getTestData returns the test data from a specified test of a specified problem
-// /problem/{id}/get/testData
-// URL params:
-//  - id - the test id
-//  - noIn - if not empty, the input file won't be sent
-//  - noOut - if not empty, the output file won't be sent
-func (s *API) getTestData(w http.ResponseWriter, r *http.Request) {
-	sid := r.FormValue("id")
-	if sid == "" {
-		errorData(w, "You must specify a test ID", http.StatusBadRequest)
-		return
-	}
-	id, err := strconv.Atoi(sid)
-	if err != nil {
-		errorData(w, "Invalid test ID", http.StatusBadRequest)
-		return
-	}
-
-	if _, err := s.db.Test(r.Context(), util.Problem(r).ID, id); err != nil {
-		errorData(w, "Test doesn't exist", 400)
-		return
-	}
-
-	var ret struct {
-		In  string `json:"in"`
-		Out string `json:"out"`
-	}
-	if r.FormValue("noIn") == "" {
-		in, err := s.manager.TestInput(id)
-		if err != nil {
-			errorData(w, err, 500)
-			return
-		}
-
-		inText, err := io.ReadAll(in)
-		if err != nil {
-			errorData(w, err, 500)
-			return
-		}
-
-		in.Close()
-		ret.In = string(inText)
-	}
-	if r.FormValue("noOut") == "" {
-		out, err := s.manager.TestOutput(id)
-		if err != nil {
-			errorData(w, err, 500)
-			return
-		}
-
-		outText, err := io.ReadAll(out)
-		if err != nil {
-			errorData(w, err, 500)
-			return
-		}
-
-		out.Close()
-		ret.Out = string(outText)
-	}
-	returnData(w, ret)
-}
-
 func (s *API) updateProblem(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm()
-	var args struct {
-		Title        *string `json:"title"`
-		Description  *string `json:"description"`
-		ConsoleInput *bool   `json:"console_input"`
-		TestName     *string `json:"test_name"`
-
-		Type kilonova.ProblemType `json:"type"`
-
-		SourceCredits *string `json:"source_credits"`
-		AuthorCredits *string `json:"author_credits"`
-
-		TimeLimit   *float64 `json:"time_limit"`
-		MemoryLimit *int     `json:"memory_limit"`
-		StackLimit  *int     `json:"stack_limit"`
-
-		DefaultPoints *int `json:"default_points"`
-
-		Visible *bool `json:"visible"`
-	}
+	var args kilonova.ProblemUpdate
 	if err := decoder.Decode(&args, r.Form); err != nil {
 		errorData(w, err, 400)
 		return
 	}
 
-	if args.Title != nil && *args.Title == "" {
-		errorData(w, "Title can't be empty", 400)
-		return
-	}
-
-	if args.Visible != nil && !util.User(r).Admin && *args.Visible != util.Problem(r).Visible {
-		errorData(w, "You can't update visibility!", 403)
-		return
-	}
-
-	if err := s.db.UpdateProblem(r.Context(), util.Problem(r).ID, kilonova.ProblemUpdate{
-		Name:         args.Title,
-		Description:  args.Description,
-		ConsoleInput: args.ConsoleInput,
-		TestName:     args.TestName,
-
-		Type: args.Type,
-
-		SourceCredits: args.SourceCredits,
-		AuthorCredits: args.AuthorCredits,
-
-		TimeLimit:   args.TimeLimit,
-		MemoryLimit: args.MemoryLimit,
-		StackLimit:  args.StackLimit,
-
-		DefaultPoints: args.DefaultPoints,
-		Visible:       args.Visible,
-	}); err != nil {
+	if err := s.base.UpdateProblem(r.Context(), util.Problem(r).ID, args, util.UserBrief(r)); err != nil {
 		errorData(w, err, 500)
 		return
 	}
