@@ -1,9 +1,8 @@
 package api
 
 import (
+	"context"
 	"net/http"
-	"strconv"
-	"strings"
 
 	"github.com/KiloProjects/kilonova"
 	"github.com/KiloProjects/kilonova/internal/util"
@@ -16,7 +15,7 @@ func (s *API) getProblemList(w http.ResponseWriter, r *http.Request) {
 		errorData(w, err, 400)
 		return
 	}
-	list, err := s.db.ProblemList(r.Context(), args.ID)
+	list, err := s.base.ProblemList(r.Context(), args.ID)
 	if err != nil {
 		errorData(w, err, 500)
 		return
@@ -24,7 +23,7 @@ func (s *API) getProblemList(w http.ResponseWriter, r *http.Request) {
 	returnData(w, list)
 }
 
-func (s *API) filterProblemList(w http.ResponseWriter, r *http.Request) {
+func (s *API) problemLists(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm()
 	var args kilonova.ProblemListFilter
 	if err := decoder.Decode(&args, r.Form); err != nil {
@@ -32,7 +31,7 @@ func (s *API) filterProblemList(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	lists, err := s.db.ProblemLists(r.Context(), args)
+	lists, err := s.base.ProblemLists(r.Context(), args)
 	if err != nil {
 		errorData(w, err, 500)
 		return
@@ -41,42 +40,33 @@ func (s *API) filterProblemList(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *API) initProblemList(w http.ResponseWriter, r *http.Request) {
-	title := r.FormValue("title")
-	description := r.FormValue("description")
-	problemIDs := r.FormValue("ids")
-	inputIDs, ok := DecodeIntString(problemIDs)
-	if !ok || len(inputIDs) == 0 {
-		errorData(w, "Invalid id list", 400)
+	var listData struct {
+		Title       string `json:"title"`
+		Description string `json:"description"`
+		ProblemIDs  []int  `json:"ids"`
+	}
+	if err := parseJsonBody(r, &listData); err != nil {
+		err.WriteError(w)
 		return
 	}
 
-	var pbs []*kilonova.Problem
-	var err error
-	if util.User(r).Admin {
-		pbs, err = s.db.Problems(r.Context(), kilonova.ProblemFilter{IDs: inputIDs})
-	} else {
-		pbs, err = s.db.Problems(r.Context(), kilonova.ProblemFilter{IDs: inputIDs, LookingUserID: &util.User(r).ID})
+	if listData.Title == "" || len(listData.ProblemIDs) == 0 {
+		errorData(w, "Invalid problem list", 400)
+		return
 	}
+
+	actualIDs, err := s.filterProblems(r.Context(), listData.ProblemIDs, util.UserBrief(r))
 	if err != nil {
-		errorData(w, err, 500)
-		return
-	}
-
-	actualIDs := make([]int, 0, len(pbs))
-	for _, pb := range pbs {
-		actualIDs = append(actualIDs, pb.ID)
-	}
-	if len(actualIDs) == 0 {
-		errorData(w, "Number of problems specified that you can see is 0", 400)
+		err.WriteError(w)
 		return
 	}
 
 	var list kilonova.ProblemList
-	list.Title = title
-	list.Description = description
-	list.AuthorID = util.User(r).ID
+	list.Title = listData.Title
+	list.Description = listData.Description
+	list.AuthorID = util.UserBrief(r).ID
 	list.List = actualIDs
-	if err := s.db.CreateProblemList(r.Context(), &list); err != nil {
+	if err := s.base.CreateProblemList(r.Context(), &list); err != nil {
 		errorData(w, err, 500)
 		return
 	}
@@ -85,50 +75,48 @@ func (s *API) initProblemList(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *API) updateProblemList(w http.ResponseWriter, r *http.Request) {
+	// TODO: Nu merge, trebuie schimbat la json body thing
 	r.ParseForm()
 	var args struct {
 		ID          int     `json:"id"`
 		Title       *string `json:"title"`
 		Description *string `json:"description"`
-		List        *string `json:"list"`
+		List        []int   `json:"list"`
 	}
 	if err := decoder.Decode(&args, r.Form); err != nil {
 		errorData(w, err, 400)
 		return
 	}
-	list, err := s.db.ProblemList(r.Context(), args.ID)
+	list, err := s.base.ProblemList(r.Context(), args.ID)
 	if err != nil {
 		errorData(w, "Couldn't find problem list", 400)
 		return
 	}
 
-	if !(util.User(r).Admin || util.User(r).ID == list.AuthorID) {
+	if !(util.UserBrief(r).Admin || util.UserBrief(r).ID == list.AuthorID) {
 		errorData(w, "You can't update this problem list!", 403)
 		return
 	}
 
-	upd := kilonova.ProblemListUpdate{
+	if err := s.base.UpdateProblemList(r.Context(), args.ID, kilonova.ProblemListUpdate{
 		Title:       args.Title,
 		Description: args.Description,
-	}
-
-	if args.List != nil {
-		ll := strings.Split(*args.List, ",")
-		l := make([]int, 0, len(ll))
-		for _, str := range ll {
-			val, err := strconv.Atoi(str)
-			if err != nil {
-				errorData(w, "Bad problem id list", 400)
-				return
-			}
-			l = append(l, val)
-		}
-		upd.List = l
-	}
-
-	if err := s.db.UpdateProblemList(r.Context(), args.ID, upd); err != nil {
+	}); err != nil {
 		errorData(w, err, 500)
 		return
+	}
+
+	if len(args.List) > 0 {
+		list, err := s.filterProblems(r.Context(), args.List, util.UserBrief(r))
+		if err != nil {
+			err.WriteError(w)
+			return
+		}
+
+		if err := s.base.UpdateProblemListProblems(r.Context(), args.ID, list); err != nil {
+			errorData(w, err, 500)
+			return
+		}
 	}
 
 	returnData(w, "Updated problem list")
@@ -141,20 +129,35 @@ func (s *API) deleteProblemList(w http.ResponseWriter, r *http.Request) {
 		errorData(w, err, 400)
 		return
 	}
-	list, err := s.db.ProblemList(r.Context(), args.ID)
+	list, err := s.base.ProblemList(r.Context(), args.ID)
 	if err != nil {
 		errorData(w, "Couldn't find problem list", 400)
 		return
 	}
 
-	if !(util.User(r).Admin || util.User(r).ID == list.AuthorID) {
+	if !(util.UserBrief(r).Admin || util.UserBrief(r).ID == list.AuthorID) {
 		errorData(w, "You can't delete this problem list!", 403)
 		return
 	}
 
-	if err := s.db.DeleteProblemList(r.Context(), args.ID); err != nil {
+	if err := s.base.DeleteProblemList(r.Context(), args.ID); err != nil {
 		errorData(w, err, 500)
 		return
 	}
 	returnData(w, "Removed problem list")
+}
+
+func (s *API) filterProblems(ctx context.Context, problemIDs []int, user *kilonova.UserBrief) ([]int, *kilonova.StatusError) {
+	if len(problemIDs) == 0 {
+		return []int{}, nil
+	}
+	pbs, err := s.base.Problems(ctx, kilonova.ProblemFilter{IDs: problemIDs, LookingUser: user})
+	if err != nil {
+		return nil, err
+	}
+	actualIDs := make([]int, 0, len(pbs))
+	for _, pb := range pbs {
+		actualIDs = append(actualIDs, pb.ID)
+	}
+	return actualIDs, nil
 }
