@@ -4,24 +4,24 @@ import (
 	"archive/zip"
 	"bufio"
 	"context"
-	"errors"
 	"fmt"
-	"io"
 	"log"
 	"path"
+	"strconv"
 	"strings"
 
 	"github.com/KiloProjects/kilonova"
+	"github.com/KiloProjects/kilonova/sudoapi"
 )
 
 var (
-	ErrBadTestFile = &kilonova.Error{Code: kilonova.EINVALID, Message: "Bad test score file"}
-	ErrBadArchive  = &kilonova.Error{Code: kilonova.EINVALID, Message: "Bad archive"}
+	ErrBadTestFile = kilonova.Statusf(400, "Bad test score file")
+	ErrBadArchive  = kilonova.Statusf(400, "Bad archive")
 )
 
 type archiveTest struct {
-	InFile  io.Reader
-	OutFile io.Reader
+	InFile  *zip.File
+	OutFile *zip.File
 	Score   int
 }
 
@@ -35,8 +35,14 @@ func NewArchiveCtx() *ArchiveCtx {
 	return &ArchiveCtx{tests: make(map[int]archiveTest), scoredTests: make([]int, 0, 10), hasScoreFile: false}
 }
 
-func ProcessArchiveFile(ctx *ArchiveCtx, name string, file io.Reader) error {
+func ProcessArchiveFile(ctx *ArchiveCtx, name string, file *zip.File) error {
 	if strings.HasSuffix(name, ".txt") { // test score file
+
+		f, err := file.Open()
+		if err != nil {
+			return kilonova.Statusf(500, "unknown error")
+		}
+		defer f.Close()
 
 		// If there's multiple score files, quit
 		if ctx.hasScoreFile {
@@ -44,7 +50,7 @@ func ProcessArchiveFile(ctx *ArchiveCtx, name string, file io.Reader) error {
 		}
 		ctx.hasScoreFile = true
 
-		br := bufio.NewScanner(file)
+		br := bufio.NewScanner(f)
 
 		for br.Scan() {
 			line := br.Text()
@@ -79,23 +85,33 @@ func ProcessArchiveFile(ctx *ArchiveCtx, name string, file io.Reader) error {
 
 	var tid int
 	if _, err := fmt.Sscanf(name, "%d-", &tid); err != nil {
-		log.Println("Bad name:", name)
-		return ErrBadArchive
+		// maybe it's problem_name.%d.{in,sol,out} format
+		nm := strings.Split(strings.TrimSuffix(name, path.Ext(name)), ".")
+		if len(nm) == 0 {
+			log.Println("Bad name:", name)
+			return ErrBadArchive
+		}
+		val, err := strconv.Atoi(nm[len(nm)-1])
+		if err != nil {
+			log.Println("Not number:", name)
+			return ErrBadArchive
+		}
+		tid = val
 	}
 
 	if strings.HasSuffix(name, ".in") { // test input file
 		tf := ctx.tests[tid]
 		if tf.InFile != nil { // in file already exists
-			return fmt.Errorf("Multiple input files for test %d", tid)
+			return fmt.Errorf("multiple input files for test %d", tid)
 		}
 
 		tf.InFile = file
 		ctx.tests[tid] = tf
 	}
-	if strings.HasSuffix(name, ".out") || strings.HasSuffix(name, ".ok") { // test output file
+	if strings.HasSuffix(name, ".out") || strings.HasSuffix(name, ".ok") || strings.HasSuffix(name, ".sol") { // test output file
 		tf := ctx.tests[tid]
 		if tf.OutFile != nil { // out file already exists
-			return fmt.Errorf("Multiple output files for test %d", tid)
+			return fmt.Errorf("multiple output files for test %d", tid)
 		}
 
 		tf.OutFile = file
@@ -104,7 +120,7 @@ func ProcessArchiveFile(ctx *ArchiveCtx, name string, file io.Reader) error {
 	return nil
 }
 
-func ProcessZipTestArchive(pb *kilonova.Problem, ar *zip.Reader, db kilonova.DB, dm kilonova.DataStore) error {
+func ProcessZipTestArchive(pb *kilonova.Problem, ar *zip.Reader, base *sudoapi.BaseAPI) error {
 	ctx := NewArchiveCtx()
 
 	for _, file := range ar.File {
@@ -112,35 +128,42 @@ func ProcessZipTestArchive(pb *kilonova.Problem, ar *zip.Reader, db kilonova.DB,
 			continue
 		}
 
-		f, err := file.Open()
-		if err != nil {
-			log.Println(err)
-			return errors.New("Unknown error")
-		}
-		defer f.Close() // This will always close all files, regardless of when the program leaves
-		if err := ProcessArchiveFile(ctx, path.Base(file.Name), f); err != nil {
+		if err := ProcessArchiveFile(ctx, path.Base(file.Name), file); err != nil {
 			return err
 		}
 	}
 
-	if len(ctx.scoredTests) != len(ctx.tests) {
+	if ctx.hasScoreFile && len(ctx.scoredTests) != len(ctx.tests) {
 		log.Println(len(ctx.scoredTests), len(ctx.tests))
-		return errors.New("Mismatched number of tests in archive and scored tests")
+		return kilonova.Statusf(400, "mismatched number of tests in archive and scored tests")
 	}
 
 	for k, v := range ctx.tests {
 		if v.InFile == nil || v.OutFile == nil {
-			return fmt.Errorf("Missing input or output file for test %d", k)
+			return kilonova.Statusf(400, "missing input or output file for test %d", k)
 		}
 	}
 
 	if !ctx.hasScoreFile {
-		return errors.New("Missing test score file")
+		log.Println("Automatically inserting scores...")
+		n := len(ctx.tests)
+		perTest := 100/n + 1
+		toSub := n - 100%n
+		k := 0
+		for i := range ctx.tests {
+			tst := ctx.tests[i]
+			tst.Score = perTest
+			if k < toSub {
+				tst.Score--
+			}
+			ctx.tests[i] = tst
+			k++
+		}
 	}
 
 	// If we are loading an archive, the user might want to remove all tests first
 	// So let's do it for them
-	if err := db.OrphanProblemTests(context.Background(), pb.ID); err != nil {
+	if err := base.OrphanTests(context.Background(), pb.ID); err != nil {
 		log.Println(err)
 		return err
 	}
@@ -150,18 +173,32 @@ func ProcessZipTestArchive(pb *kilonova.Problem, ar *zip.Reader, db kilonova.DB,
 		test.ProblemID = pb.ID
 		test.VisibleID = testID
 		test.Score = v.Score
-		if err := db.CreateTest(context.Background(), &test); err != nil {
+		if err := base.CreateTest(context.Background(), &test); err != nil {
 			log.Println(err)
 			return err
 		}
 
-		if err := dm.SaveTestInput(test.ID, v.InFile); err != nil {
-			log.Println("Couldn't create test input", err)
-			return fmt.Errorf("Couldn't create test input: %w", err)
+		{
+			f, err := v.InFile.Open()
+			if err != nil {
+				return kilonova.WrapError(err, "Couldn't open() input file")
+			}
+			defer f.Close()
+			if err := base.SaveTestInput(test.ID, f); err != nil {
+				log.Println("Couldn't create test input", err)
+				return kilonova.WrapError(err, "Couldn't create test input")
+			}
 		}
-		if err := dm.SaveTestOutput(test.ID, v.OutFile); err != nil {
-			log.Println("Couldn't create test output", err)
-			return fmt.Errorf("Couldn't create test output: %w", err)
+		{
+			f, err := v.OutFile.Open()
+			if err != nil {
+				return kilonova.WrapError(err, "Couldn't open() output file")
+			}
+			defer f.Close()
+			if err := base.SaveTestOutput(test.ID, f); err != nil {
+				log.Println("Couldn't create test output", err)
+				return kilonova.WrapError(err, "Couldn't create test output")
+			}
 		}
 	}
 
