@@ -8,15 +8,15 @@ import (
 	"strconv"
 )
 
-func (m *StorageManager) TestInput(testID int) (io.ReadCloser, error) {
+func (m *StorageManager) TestInput(testID int) (io.ReadSeekCloser, error) {
 	return os.Open(path.Join(m.RootPath, "tests", strconv.Itoa(testID)+".in"))
 }
 
-func (m *StorageManager) TestOutput(testID int) (io.ReadCloser, error) {
-	return os.Open(m.TestOutputPath(testID))
+func (m *StorageManager) TestOutput(testID int) (io.ReadSeekCloser, error) {
+	return os.Open(m.testOutputPath(testID))
 }
 
-func (m *StorageManager) TestOutputPath(testID int) string {
+func (m *StorageManager) testOutputPath(testID int) string {
 	return path.Join(m.RootPath, "tests", strconv.Itoa(testID)+".out")
 }
 
@@ -25,5 +25,5 @@ func (m *StorageManager) SaveTestInput(testID int, input io.Reader) error {
 }
 
 func (m *StorageManager) SaveTestOutput(testID int, output io.Reader) error {
-	return writeFile(path.Join(m.RootPath, "tests", fmt.Sprintf("%d.out", testID)), output, 0777)
+	return writeFile(m.testOutputPath(testID), output, 0777)
 }
