@@ -13,26 +13,16 @@ import (
 var (
 	configPath string
 	Common     CommonConf
-	Database   DBConf
 	Eval       EvalConf
 	Email      EmailConf
-	Index      IndexConf
 )
 
 // configStruct is the glue for all configuration sections when unmarshaling
 // After load, it will disperse all the data in variables
 type configStruct struct {
-	Common   CommonConf `toml:"common"`
-	Database DBConf     `toml:"database"`
-	Eval     EvalConf   `toml:"eval"`
-	Email    EmailConf  `toml:"email"`
-	Index    IndexConf  `toml:"index"`
-}
-
-type IndexConf struct {
-	Lists        []int  `toml:"lists_to_show"`
-	ShowProblems bool   `toml:"show_problems"`
-	Description  string `toml:"description"`
+	Common CommonConf `toml:"common"`
+	Eval   EvalConf   `toml:"eval"`
+	Email  EmailConf  `toml:"email"`
 }
 
 // EmailConf is the data required for the email part
@@ -40,6 +30,7 @@ type EmailConf struct {
 	Host     string `toml:"host"`
 	Username string `toml:"username"`
 	Password string `toml:"password"`
+	SendAs   string `toml:"sendAs"`
 }
 
 // EvalConf is the data required for the eval service
@@ -58,12 +49,8 @@ type CommonConf struct {
 	HostPrefix  string `toml:"host_prefix"`
 	Port        int    `toml:"port"`
 	DefaultLang string `toml:"default_language"`
-}
 
-// DBConf is the data required to establish a PostgreSQL connection
-type DBConf struct {
-	Type string `toml:"dbtype"`
-	DSN  string `toml:"dsn"`
+	DBDSN string `toml:"db_dsn"`
 }
 
 // c represents the loaded config
@@ -71,18 +58,14 @@ var c configStruct
 
 func spread() {
 	Common = c.Common
-	Database = c.Database
 	Email = c.Email
 	Eval = c.Eval
-	Index = c.Index
 }
 
 func compactify() {
 	c.Common = Common
-	c.Database = Database
 	c.Email = Email
 	c.Eval = Eval
-	c.Index = Index
 }
 
 func SetConfigPath(path string) {
@@ -92,7 +75,7 @@ func SetConfigPath(path string) {
 func Save() error {
 	compactify()
 	if configPath == "" {
-		return errors.New("Invalid config path")
+		return errors.New("invalid config path")
 	}
 
 	// Make the directories just in case they don't exist
@@ -117,7 +100,7 @@ func Save() error {
 
 func Load() error {
 	if configPath == "" {
-		return errors.New("Invalid config path")
+		return errors.New("invalid config path")
 	}
 	md, err := toml.DecodeFile(configPath, &c)
 	if len(md.Undecoded()) > 0 {
