@@ -4,6 +4,7 @@ import (
 	"bytes"
 
 	mathjax "github.com/litao91/goldmark-mathjax"
+	"github.com/microcosm-cc/bluemonday"
 	"github.com/yuin/goldmark"
 	"github.com/yuin/goldmark/extension"
 	"github.com/yuin/goldmark/parser"
@@ -12,13 +13,14 @@ import (
 
 // LocalRenderer is a local markdown renderer. It does not depend on any external services but it does not support mathjax rendering or any extensions to the markdown standard I intend to make.
 type LocalRenderer struct {
-	md goldmark.Markdown
+	md  goldmark.Markdown
+	pol *bluemonday.Policy
 }
 
 func (r *LocalRenderer) Render(src []byte) ([]byte, error) {
 	var buf bytes.Buffer
 	err := r.md.Convert(src, &buf)
-	return buf.Bytes(), err
+	return r.pol.SanitizeReader(&buf).Bytes(), err
 }
 
 func NewLocalRenderer() *LocalRenderer {
@@ -27,5 +29,6 @@ func NewLocalRenderer() *LocalRenderer {
 		goldmark.WithParserOptions(parser.WithAutoHeadingID(), parser.WithAttribute()),
 		goldmark.WithRendererOptions(html.WithHardWraps(), html.WithXHTML()),
 	)
-	return &LocalRenderer{md}
+	pol := bluemonday.UGCPolicy()
+	return &LocalRenderer{md, pol}
 }
